+++ Frameworks/Adium Framework/Source/AIListObject.m
@@ -248,8 +248,8 @@
 		 * Dropping between Away Contact and Offline Contact should make an Away Contact be > 120 but an Offline Contact be < 110.
 		 * Only the sort controller knows the answer as to where this contact should be positioned in the end.
 		 */
-		AILogWithSignature(@"%@: Moving %@ into %@'s index %i using order index %f (between %@ and %@)",
-						   container, listObject, container.visibleContainedObjects, idx, 
+		AILogWithSignature(@"%@: Moving %@ into %@'s index %li using order index %f (between %@ and %@)",
+						   container, listObject, container.visibleContainedObjects, (long)idx, 
 						   (nextHighest + nextLowest) / 2, nextObject, previousObject);
 		[container listObject: listObject didSetOrderIndex: (nextHighest + nextLowest) / 2];
 	}
